requirements.txt
python-telegram-bot==20.3
flask==2.3.2
python-dotenv==1.0.0
gunicorn==20.1.0